requirements.txt
google-genai
google-adk
pypdf
streamlit
